Create a Solver class object using an ensemble of solvers

Create a Solver class object using an ensemble of solvers

EnsembleSolver(
  mtx.assay = matrix(),
  targetGene,
  candidateRegulators,
  solverNames = c("lasso", "lassopv", "pearson", "randomForest", "ridge", "spearman",
    "xgboost"),
  geneCutoff = 0.1,
  alpha.lasso = 0.9,
  alpha.ridge = 0,
  lambda.lasso = numeric(0),
  lambda.ridge = numeric(0),
  lambda.sqrt = numeric(0),
  nCores.sqrt = 4,
  nOrderings.bayes = 10,
  quiet = TRUE
)

Arguments

mtx.assay
An assay matrix of gene expression data

targetGene
A designated target gene that should be part of the mtx.assay data

candidateRegulators
The designated set of transcription factors that could be associated with the target gene

solverNames
A character vector of strings denoting

geneCutoff
A fraction (0-1) of the supplied candidate regulators to be included in the fetaures output by the solver (default = 0.1)

alpha.lasso
A fraction (0-1) denoting the LASSO-Ridge balance of the `glmnet` solver used by the LASSO method (default = 0.9)

alpha.ridge
A fraction (0-1) denoting the LASSO-Ridge balance of the `glmnet` solver used by the Ridge method (default = 0)

lambda.lasso
The penalty parameter for LASSO, used to determine how strictly to penalize the regression coefficients. If none is supplied, this will be determined via permutation testing (default = NULL).

lambda.ridge
The penalty parameter for Ridge, used to determine how strictly to penalize the regression coefficients. If none is supplied, this will be determined via permutation testing (default = NULL).

lambda.sqrt
The penalty parameter for square root LASSO, used to determine how strictly to penalize the regression coefficients. If none is supplied, this will be determined via permutation testing (default = NULL).

nCores.sqrt
An integer denoting the number of computational cores to devote to the square root LASSO solver, which is the slowest of the solvers (default = 4)

nOrderings.bayes
An integer denoting the number of random starts to use for the Bayes Spike method (default = 10)

quiet
A logical denoting whether or not the solver should print output

Value

A Solver class object with Ensemble as the solver

See also

Examples

load(system.file(package="trena", "extdata/ampAD.154genes.mef2cTFs.278samples.RData"))
target.gene <- "MEF2C"
tfs <- setdiff(rownames(mtx.sub), target.gene)
ensemble.solver <- EnsembleSolver(mtx.sub, target.gene, tfs)