Class FootprintFinder

The FootprintFinder class is designed to query 2 supplied footprint databases (a genome database and a project database) for supplied genes or regions. Within the TReNA package, the FootprintFinder class is mainly used by the FootprintFilter class, but the FootprintFinder class offers more flexibility in constructing queries.

FootprintFinder(genome.database.uri, project.database.uri, quiet = TRUE)

Arguments

genome.database.uri
The address of a genome database for use in filtering. This database must contain the tables "gtf" and "motifsgenes" at a minimum. The URI format is as follows: "dbtype://host/database" (e.g. "postgres://localhost/genomedb")

project.database.uri
The address of a project database for use in filtering. This database must contain the tables "regions" and "hits" at a minimum. The URI format is as follows: "dbtype://host/database" (e.g. "postgres://localhost/projectdb")

quiet
A logical denoting whether or not the FootprintFinder object should print output

Value

An object of the FootprintFinder class

Slots

genome.db

The address of a genome database for use in filtering

project.db

The address of a project database for use in filtering

quiet

A logical argument denoting whether the FootprintFinder object should behave quietly

See also