Create a HumanDHSFilter object

A HumanDHSFilter object allows for filtering based on DNAse hypersensitivity (DHS) data. Its associated getCandidates method uses a genome from a BSgenome database (either hg19 or hg38), DNA region specifications, and (variants/pfms,encodetablename, match to filter a list of possible regulators factors to those that match the supplied criteria.

HumanDHSFilter(genomeName, encodeTableName = "wgEncodeRegDnaseClustered",
  pwmMatchPercentageThreshold, geneInfoDatabase.uri, regions,
  variants = NA_character_, pfms, quiet = TRUE)

Arguments

genomeName
A character string indicating the reference genome; currently, the only accepted strings are "hg38" and "hg19", both of which are human genomes.

encodeTableName
(default = "wgEncodeRegDnaseClustered")

pwmMatchPercentageThreshold
A numeric from 0-100 to serve as a threshold for a match

geneInfoDatabase.uri
An address for a gene database

regions
A data frame containing the regions of interest

variants
A character vector containing a list of variants

pfms
A list of position frequency matrices, often converted from a MotifList object created by a MotifDb query

quiet
A logical denoting whether or not the solver should print output

Value

A CandidateFilter class object that filters using Human DHS data

See also

Examples

if(interactive()) { # takes too long in the bioc windows build
  load(system.file(package="trena", "extdata/ampAD.154genes.mef2cTFs.278samples.RData"))
  targetGene <- "VRK2"
  promoter.length <- 1000
  genomeName <- "hg38"
  db.address <- system.file(package="trena", "extdata")
  genome.db.uri    <- paste("sqlite:/", db.address, "vrk2.neighborhood.hg38.gtfAnnotation.db",  sep = "/")

    # Grab regions for VRK2 using shoulder size of 1000
  trena <- Trena(genomeName)
  tbl.regions <- getProximalPromoter(trena, "VRK2", 1000, 1000)

  hd.filter <- HumanDHSFilter(genomeName, pwmMatchPercentageThreshold = 85,
  geneInfoDatabase.uri = genome.db.uri, regions = tbl.regions,
   pfms = as.list(query(query(MotifDb, "sapiens"),"jaspar2016")))
   } # if interactive