Create a Solver class object using the Square Root LASSO solver

Create a Solver class object using the Square Root LASSO solver

SqrtLassoSolver(
  mtx.assay = matrix(),
  targetGene,
  candidateRegulators,
  regulatorWeights = rep(1, length(candidateRegulators)),
  lambda = numeric(0),
  nCores = 4,
  quiet = TRUE
)

Arguments

mtx.assay
An assay matrix of gene expression data

targetGene
A designated target gene that should be part of the mtx.assay data

candidateRegulators
The designated set of transcription factors that could be associated with the target gene

regulatorWeights
A set of weights on the transcription factors (default = rep(1, length(tfs)))

lambda
A tuning parameter that determines the severity of the penalty function imposed on the elastic net regression. If unspecified, lambda will be determined via permutation testing (default = numeric(0)).

nCores
An integer specifying the number of computational cores to devote to this square root LASSO solver. This solver is generally quite slow and is greatly sped up when using multiple cores (default = 4)

quiet
A logical denoting whether or not the solver should print output

Value

A Solver class object with Square Root LASSO as the solver

See also

Examples

load(system.file(package="trena", "extdata/ampAD.154genes.mef2cTFs.278samples.RData"))
target.gene <- "MEF2C"
tfs <- setdiff(rownames(mtx.sub), target.gene)
sqrt.solver <- SqrtLassoSolver(mtx.sub, target.gene, tfs)