Get Gene Promoter Region

Using the getChromLoc function in conjunction with the gtf table inside the genome database specified by the FootprintFinder object, get the chromosome, starting location, and ending location for gene promoter region.

# S4 method for FootprintFinder
getGenePromoterRegion(obj, gene,
  size.upstream = 1000, size.downstream = 0,
  biotype = "protein_coding", moleculetype = "gene")

Arguments

obj
An object of class FootprintFinder

gene
A gene name of ID

size.upstream
An integer denoting the distance upstream of the target gene to look for footprints (default = 1000)

size.downstream
An integer denoting the distance downstream of the target gene to look for footprints (default = 0)

biotype
A type of biological unit (default="protein_coding")

moleculetype
A type of molecule (default="gene")

Value

A list containing 3 elements: 1) chr : The name of the chromosome containing the promoter region for the specified gene 2) start : The starting location of the promoter region for the specified gene 3) end : The ending location of the promoter region for the specified gene

See also

Examples

db.address <- system.file(package="trena", "extdata")
genome.db.uri <- paste("sqlite:/",db.address,"mef2c.neighborhood.hg38.gtfAnnotation.db", sep = "/")
project.db.uri <- paste("sqlite:/",db.address,"mef2c.neigborhood.hg38.footprints.db", sep = "/")
fp <- FootprintFinder(genome.db.uri, project.db.uri)

prom.region <- getGenePromoterRegion(fp, gene = "MEF2C")