Inferring Transcriptional Regulation with TReNA

`trena` provides a framework for using gene expression data to infer relationships between a target gene and a set of transcription factors. It does so using a several classes and their associated methods, briefly documented below

Details

#' Solver Class Objects

The Solver class is a base class used within `trena`. A particular Solver object also contains the name of the selected solver and dispatches the correct feature selection method when run is called on the object. It is inherited by all the following subclasses, representing the different feature selection methods: BayesSpikeSolver, EnsembleSolver, LassoPVSolver, LassoSolver, PearsonSolver, RandomForestSolver, RidgeSolver, SpearmanSolver, SqrtLassoSolver.

CandidateFilter Class Objects

The CandidateFilter class is a base class that is generally used to filter the transcription factors in the expression matrix to obtain a set of candidate regulators. Filtering method depends on the filter type chosen; there are currently the following subclasses: FootprintFilter, HumanDHSFilter, GeneOntologyFilter, and VarianceFilter. The filters are applied using the getCandidates method on a given CandidateFilter object.

FootprintFinder Class Objects

The FootprintFinder class is designed to allow extraction of gene footprinting information from existing PostgreSQL or SQLite databases. In standard use of the `trena` package, it is used solely by the getCandidates method for a FootprintFilter object. However, a FootprintFinder object has many more available methods that allow it to extract information more flexibly.